+++ package.json
@@ -6,10 +6,10 @@
   "type": "module",
   "scripts": {
     "build": "tsc",
-    "CreateCollection": "tsc && node --no-warnings dist/scripts/createItems/InscribeCollection.js",
-    "CreateItems": "tsc && node --no-warnings dist/scripts/createItems/ProcessFullItemCreationOrder.js",
-    "GetInscriptionOrder": "tsc && node --no-warnings dist/scripts/createItems/GetInscriptionOrder.js",
-    "InscribeCollectionInBatch": "tsc && node --no-warnings dist/scripts/createItems/createItemBatch/InscribeStaticCollectionItems.js",
+    "CreateCollection": "tsc && node --no-warnings dist/scripts/create/CreateCollection.js",
+    "CreateItems": "tsc && node --no-warnings dist/scripts/create/CreateItems.js",
+    "GetInscriptionOrder": "tsc && node --no-warnings dist/scripts/create/GetInscriptionOrder.js",
+    "InscribeCollectionInBatch": "tsc && node --no-warnings dist/scripts/create/createItemBatch/InscribeStaticCollectionItems.js",
     "sendAllOrderItems": "tsc && node --no-warnings dist/scripts/util/sendAllOrderItems.js",
     "createCatalog": "tsc && node --no-warnings dist/scripts/catalog/CreateCatalog.js",
     "addCollectionOrderItemsToCatalog": "tsc && node --no-warnings dist/scripts/catalog/AddCollectionOrderItemsToCatalog.js",
