+++ src/scripts/create/createItemBatch/AddStaticItemsToCreationOrder.ts
