+++ src/scripts/create/createItemBatch/InscribeStaticCollectionItems.ts
